requirements.txt
pyTelegramBotAPI
yfinance
requests
schedule
python-dotenv
flask
